leaf_spans moved from comments to byte_span
