+++ sway-fmt-v2/src/utils/byte_span.rs
@@ -1,6 +1,15 @@
 use std::cmp::Ordering;
-
-use sway_types::Span;
+use sway_ast::{
+    attribute::Annotated,
+    brackets::{Parens, SquareBrackets},
+    keywords::{
+        AddToken, ColonToken, CommaToken, ForToken, ForwardSlashToken, RightArrowToken,
+        SemicolonToken,
+    },
+    Braces, TypeField,
+};
+use sway_parse::Parse;
+use sway_types::{Ident, Span, Spanned};
 
 /// A stripped down version of sway-types::src::Span
 #[derive(PartialEq, Eq, Debug, Clone, Default)]
@@ -48,6 +57,156 @@ impl PartialOrd for ByteSpan {
 /// `leaf_spans` collects all field spans so that we can check in between them.
 pub trait LeafSpans {
     fn leaf_spans(&self) -> Vec<ByteSpan>;
+}
+
+impl<T> LeafSpans for Braces<T>
+where
+    T: LeafSpans + Clone,
+{
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        let mut collected_spans = Vec::new();
+        let mut opening_brace_span = ByteSpan::from(self.span());
+        opening_brace_span.end = opening_brace_span.start + 1;
+        // Add opening brace's ByteSpan
+        collected_spans.push(opening_brace_span);
+        // Add T's collected ByteSpan
+        collected_spans.append(&mut self.clone().into_inner().leaf_spans());
+        let mut closing_brace_span = ByteSpan::from(self.span());
+        closing_brace_span.start = closing_brace_span.end - 1;
+        // Add closing brace's ByteSpan
+        collected_spans.push(closing_brace_span);
+        collected_spans
+    }
+}
+
+impl<T> LeafSpans for Parens<T>
+where
+    T: LeafSpans + Clone,
+{
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        let mut collected_spans = Vec::new();
+        let mut opening_paren_span = ByteSpan::from(self.span());
+        opening_paren_span.end = opening_paren_span.start + 1;
+        // Add opening paren's span
+        collected_spans.push(opening_paren_span);
+        // Add T's collected ByteSpan
+        collected_spans.append(&mut self.clone().into_inner().leaf_spans());
+        let mut closing_paren_span = ByteSpan::from(self.span());
+        closing_paren_span.start = closing_paren_span.end - 1;
+        // Add closing paren's ByteSpan
+        collected_spans.push(closing_paren_span);
+        collected_spans
+    }
+}
+
+impl<T> LeafSpans for SquareBrackets<T>
+where
+    T: LeafSpans + Clone,
+{
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        let mut collected_spans = Vec::new();
+        let mut opening_bracket_span = ByteSpan::from(self.span());
+        opening_bracket_span.end = opening_bracket_span.start + 1;
+        // Add opening bracket's span
+        collected_spans.push(opening_bracket_span);
+        // Add T's collected ByteSpan
+        collected_spans.append(&mut self.clone().into_inner().leaf_spans());
+        let mut closing_bracket_span = ByteSpan::from(self.span());
+        closing_bracket_span.start = closing_bracket_span.end - 1;
+        // Add closing bracket's ByteSpan
+        collected_spans.push(closing_bracket_span);
+        collected_spans
+    }
+}
+
+impl<T, P> LeafSpans for (T, P)
+where
+    T: LeafSpans,
+    P: LeafSpans,
+{
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        let mut collected_spans = self.0.leaf_spans();
+        collected_spans.append(&mut self.1.leaf_spans());
+        collected_spans
+    }
+}
+impl<T> LeafSpans for Vec<T>
+where
+    T: LeafSpans,
+{
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        let mut collected_spans = Vec::new();
+        for t in self {
+            collected_spans.append(&mut t.leaf_spans());
+        }
+        collected_spans
+    }
+}
+impl<T> LeafSpans for Annotated<T>
+where
+    T: LeafSpans + Parse,
+{
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        let mut collected_spans = self.attribute_list.leaf_spans();
+        collected_spans.append(&mut self.value.leaf_spans());
+        collected_spans
+    }
+}
+
+impl LeafSpans for Ident {
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        vec![ByteSpan::from(self.span())]
+    }
+}
+impl LeafSpans for CommaToken {
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        vec![(ByteSpan::from(self.span()))]
+    }
+}
+
+impl LeafSpans for TypeField {
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        let mut collected_spans = vec![ByteSpan::from(self.name.span())];
+        collected_spans.push(ByteSpan::from(self.colon_token.span()));
+        collected_spans.append(&mut self.ty.leaf_spans());
+        collected_spans
+    }
+}
+
+impl LeafSpans for AddToken {
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        vec![ByteSpan::from(self.span())]
+    }
+}
+
+impl LeafSpans for SemicolonToken {
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        vec![ByteSpan::from(self.span())]
+    }
+}
+
+impl LeafSpans for ColonToken {
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        vec![ByteSpan::from(self.span())]
+    }
+}
+
+impl LeafSpans for RightArrowToken {
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        vec![ByteSpan::from(self.span())]
+    }
+}
+
+impl LeafSpans for ForToken {
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        vec![ByteSpan::from(self.span())]
+    }
+}
+
+impl LeafSpans for ForwardSlashToken {
+    fn leaf_spans(&self) -> Vec<ByteSpan> {
+        vec![ByteSpan::from(self.span())]
+    }
 }
 
 #[cfg(test)]
